Initial version of Scala scripting provider
Allows script files with .scala extension
under $GERRIT_SITE/plugins to be loaded as ServerPlugins.

Manifest is automatically generated on-the-fly based
on the Guice Module classes defined in the Scala script.
When no Guice Modules are defined, classes are automatically
registered as in standard Jar plugins.

At the moment all Scala scripts are defined as PLUGIN
type: it is not yet possible to define a Gerrit EXTENSION type.

All Scala script plugins are defined as hot-reloadable.

Change-Id: Id7ad0b51bf07b57b093ddf3148165bdfc51fd685