Allows multiple Scala scripts to be included in a single plugin
When the 'plugin-version.scala' under /plugins is a directory 
all the Scala scripts inside the directory tree are loaded
inside the same plugin environment.

This allows to save precious PermGen JVM Heap and reuse
the same plugin name space for multiple Scala scripts.
Additionally allows to have a much better organisation of
larger set of scripts in packages.

Change-Id: Ic885d4c8d89ce93f801fc642869e21378abe4fd0